Change the display PHP version for the bundled PHP
Rather than adding the bundled version to the list used to generate the
version lines, instead append an extra line on the end. This allows for
passing of the version number, while still being able to check if it's
in use - the previous method of checking the `php.default` would not
longer work, as the version index is now the version number instead of
`bundled`. Instead, manually check the `php.default` against `bundled`
for this additional line.

Fixes #160